+++ partner-chains-cli/src/main.rs
@@ -2,6 +2,7 @@ mod cardano_cli;
 pub mod config;
 pub mod create_chain_spec;
 mod dereg;
+mod deregister;
 pub mod generate_keys;
 pub mod io;
 pub mod keystore;
@@ -45,6 +46,8 @@ pub enum Command {
 	/// The final step of registering as a committee candidate, not using cold keys.
 	Register3(register::register3::Register3Cmd),
 	Dereg(dereg::DeregCmd),
+	/// Deregister from the candidates set. This command requires chain config file present in the running directory.
+	Deregister(deregister::DeregisterCmd),
 }
 
 pub trait CmdRun {
@@ -63,6 +66,7 @@ fn main() -> anyhow::Result<()> {
 		Command::Register2(cmd) => cmd.run(&DefaultCmdRunContext)?,
 		Command::Register3(cmd) => cmd.run(&DefaultCmdRunContext)?,
 		Command::Dereg(cmd) => cmd.run(&DefaultCmdRunContext)?,
+		Command::Deregister(cmd) => cmd.run(&DefaultCmdRunContext)?,
 	}
 	Ok(())
 }
@@ -88,6 +92,7 @@ const HELP_EXAMPLES: &str = r#"
 ║   3. register2             : complete registration with cold keys              ║
 ║   4. register3             : finalize registration                             ║
 ║   5. start-node            : start the validator node                          ║
+║   6. deregister            : cancels registration                              ║
 ║                                                                                ║
 ║   Note: This sequence assumes that the chain-spec.json and                     ║
 ║         partner-chains-cli-chain-config.json files have been obtained from     ║